cleanup test TODOs
